* Lib/Console: optimize.

+++ Lib/Obj64/Console.c
@@ -144,7 +144,7 @@ void Console_Hex (INTEGER i)
 	INTEGER k, n;
 	k = -28;
 	while (k <= 0) {
-		n = (INTEGER)__MASK(__ASH(i, k, INTEGER), -16);
+		n = __MASK(__ASH(i, k, INTEGER), -16);
 		if (n <= 9) {
 			Console_Char((CHAR)(48 + n));
 		} else {
@@ -160,7 +160,7 @@ void Console_LongHex (LONGINT i)
 	INTEGER k, n;
 	k = -60;
 	while (k <= 0) {
-		n = (INTEGER)__MASK(__ASH(i, k, LONGINT), -16);
+		n = __MASK((INTEGER)__ASH(i, k, LONGINT), -16);
 		if (n <= 9) {
 			Console_Char((CHAR)(48 + n));
 		} else {
